refactor: #102 프로모션 상세 조회 API '저장 수', '내 저장 여부' 조회 로직 추가

+++ src/main/kotlin/com/damaba/damaba/mapper/PromotionMapper.kt
@@ -34,7 +34,7 @@ abstract class PromotionMapper {
     abstract fun toPromotionDetail(
         promotion: Promotion,
         author: User?,
-        saveCount: Int,
+        saveCount: Long,
         isSaved: Boolean,
     ): PromotionDetail
